+++ packages/demo/src/components/_CodeBlock/codeBlock.tsx
@@ -1,5 +1,6 @@
 import Prism from 'prismjs';
 import React, { useRef, useState, useEffect, TransitionEventHandler } from 'react';
+import './index.scss';
 interface ICodeProps {
   children: React.ReactNode;
   code: string;
