--- packages/demo/src/images/mountain.min.tsx
@@ -1,10 +1,0 @@
-export default function Mountain(props) {
-  return (
-    <svg width="100vw" height="820" xmlns="http://www.w3.org/2000/svg" version="1.0">
-      <path
-        stroke="null"
-        d="M1474.741 500.046c-7.541 2.7-13.392 7.6-32.894 26.9-9.491 9.4-23.143 22-30.424 28-7.15 5.9-18.982 16.9-26.133 24.3-7.151 7.4-20.803 19.6-30.164 27-9.361 7.4-20.933 17.4-25.743 22.1-4.811 4.7-13.522 11.7-19.373 15.5-19.502 12.7-28.213 19-32.244 23.2-2.08 2.3-4.42 4.2-5.07 4.2s-5.071 4-9.882 9c-8.19 8.5-8.97 9-13.522 9-10.4 0-24.313 2.6-56.817 10.6l-18.332 4.6-6.37-2.5c-9.362-3.6-20.153-5.1-33.025-4.3-6.11.3-19.503 1-29.904 1.6-17.292.9-25.613 2.1-45.506 6.3-9.23 2-38.615 5.7-44.855 5.7-3.25 0-6.111-.8-8.841-2.6-2.08-1.4-7.281-3.8-11.312-5.4-4.16-1.6-14.172-6.2-22.233-10.3-8.19-4.1-18.332-8.7-22.493-10.2-4.16-1.5-13.261-6-20.152-10-17.032-9.8-30.034-15.1-40.435-16.4-4.81-.7-13.912-3.1-21.063-5.7-13.652-5-22.233-6.3-41.995-6.4-11.572 0-32.764 3.5-43.166 7.1-9.23 3.3-30.423 20.2-41.345 33-4.94 5.7-14.042 14.5-20.152 19.5-6.241 5-15.732 13.5-21.063 19.1-11.832 11.9-19.503 17.3-30.424 21.3-10.271 3.8-17.162 3.9-25.873.4-32.634-13.4-63.968-31.8-102.713-60.3-7.931-5.8-21.713-15-30.814-20.4-31.464-18.7-46.416-33.3-67.479-65.8-12.871-19.8-25.873-34.8-40.695-46.8-26.393-21.4-34.324-25.3-56.427-27.7-7.671-.9-17.162-2.5-21.063-3.6-26.783-7.5-24.573-7.1-33.804-5.9-11.831 1.5-25.743 7.1-35.884 14.3-4.551 3.3-16.643 14-26.784 23.7-29.774 28.3-56.557 47.7-82.43 59.6-12.092 5.6-17.032 8.6-24.443 14.8-27.434 23.3-32.765 26.9-55.647 38-14.432 6.9-20.153 8.2-34.325 7.3-16.252-.9-25.093-4.6-47.326-19.5-17.682-11.9-26.263-18.5-34.324-26.6-4.29-4.3-8.581-8.6-9.491-9.5-1.3-1.2-1.56 73.6-1.69 350.2v351.8h1664.21v-775.8l-6.76-3.9c-3.771-2.2-10.402-5.7-14.693-7.9-4.29-2.2-15.081-7.8-24.052-12.5-12.742-6.7-20.023-9.7-33.545-13.8-15.862-4.8-18.332-5.3-30.554-5.7-11.181-.4-14.041-.2-18.462 1.4z"
-      />
-    </svg>
-  );
-}
