+++ packages/demo/src/images/mountains/mountain.min.tsx
@@ -1,0 +1,10 @@
+export default function Mountain(props) {
+  return (
+    <svg width="1980" height="820" {...props} xmlns="http://www.w3.org/2000/svg" version="1.0">
+      <path
+        stroke="null"
+        d="M1826.956 453.966c-9.27 2.858-16.462 8.044-40.436 28.473-11.667 9.95-28.45 23.286-37.4 29.637-8.789 6.245-23.334 17.888-32.125 25.72-8.79 7.833-25.572 20.746-37.08 28.579-11.507 7.832-25.732 18.417-31.645 23.392-5.914 4.975-16.622 12.384-23.815 16.406-23.974 13.442-34.682 20.11-39.637 24.556-2.557 2.435-5.434 4.446-6.233 4.446s-6.233 4.234-12.147 9.526c-10.068 8.997-11.027 9.526-16.623 9.526-12.784 0-29.887 2.752-69.844 11.22l-22.535 4.869-7.83-2.646c-11.51-3.81-24.775-5.399-40.598-4.552-7.51.318-23.975 1.059-36.76 1.694-21.257.952-31.486 2.223-55.94 6.668-11.347 2.117-47.47 6.033-55.14 6.033-3.995 0-7.512-.846-10.868-2.752-2.557-1.482-8.95-4.022-13.906-5.715-5.114-1.694-17.421-6.563-27.33-10.903-10.068-4.34-22.536-9.208-27.65-10.796-5.115-1.588-16.302-6.35-24.773-10.584-20.938-10.373-36.92-15.983-49.706-17.36-5.913-.74-17.102-3.28-25.893-6.032-16.782-5.293-27.33-6.669-51.624-6.775-14.225 0-40.276 3.705-53.063 7.515-11.346 3.493-37.399 21.381-50.825 34.93-6.073 6.033-17.262 15.347-24.773 20.64-7.672 5.292-19.339 14.289-25.892 20.216-14.545 12.596-23.975 18.312-37.4 22.546-12.626 4.022-21.097 4.128-31.805.423-40.117-14.183-78.635-33.66-126.264-63.825-9.75-6.14-26.691-15.877-37.879-21.593-38.678-19.793-57.058-35.247-82.95-69.647-15.823-20.957-31.806-36.834-50.027-49.536-32.444-22.651-42.194-26.779-69.364-29.32-9.43-.952-21.097-2.645-25.893-3.81-32.924-7.938-30.207-7.515-41.555-6.245-14.543 1.588-31.645 7.515-44.111 15.136-5.595 3.493-20.46 14.819-32.926 25.086-36.6 29.954-69.524 50.489-101.33 63.084-14.864 5.928-20.937 9.103-30.047 15.666-33.724 24.662-40.277 28.472-68.406 40.221-17.741 7.304-24.774 8.68-42.195 7.727-19.979-.953-30.847-4.869-58.177-20.64-21.737-12.596-32.285-19.582-42.194-28.155a5474.713 5474.713 0 00-11.668-10.056c-1.598-1.27-1.917 77.903-2.077 370.674V1320h2045.789V498.845l-8.31-4.128c-4.636-2.329-12.787-6.033-18.062-8.362-5.273-2.329-18.539-8.256-29.566-13.23-15.664-7.093-24.615-10.268-41.237-14.608-19.499-5.08-22.535-5.61-37.56-6.033-13.744-.423-17.26-.212-22.695 1.482z"
+      />
+    </svg>
+  );
+}
