+++ packages/demo/src/images/mountains/mountain1.min.tsx
@@ -1,15 +1,25 @@
-export default function Mountain1() {
+export default function Mountain1(props) {
   return (
-    <svg width="1416" height="500" xmlns="http://www.w3.org/2000/svg">
+    <svg width="1980" height="500" xmlns="http://www.w3.org/2000/svg">
       <defs>
-        <linearGradient id="prefix__a" x1="0" y1="0" x2="0" y2="1">
-          <stop stopColor="#faf" stopOpacity=".996" offset="0" />
-          <stop stopColor="#7f007f" stopOpacity=".996" offset="1" />
+        <linearGradient y2="1" x2="0" y1="0" x1="0" id="prefix__a">
+          <stop offset="0" stopOpacity="0.996" stopColor="#faf" />
+          <stop offset="1" stopOpacity="0.996" stopColor="#7f007f" />
         </linearGradient>
       </defs>
-      <g stroke="null" fill="url(#prefix__a)">
-        <path d="M1419.473 520.321c-10.21-.37 19.736-355.006-12.958-315.008-2.368-5.678 17.856-12.97.614-11.414-7.663-.856 1.808-7.778-3.108-9.334-4.5-.816 3.361-8.42-3.976-6.591-15.796-2.625 13.374-10.85.235-11.531-2.097.272-4.808-1.614-3.886-3.636 10.157-8.673-8.169-.603-3.271-7.778 7.337-3.092 2.259-5.037-2.801-4.9-.778-.428.921-13.281-5.17-7.778 2.675 14.408-11.494 4.977-4.066 12.755 9.29 5.834-9.036-.486-.144 9.84-14.115-.7 3.253 6.339-2.404 10.383-9.38-1.128 1.392 4.784-.524 7.875-26.35-5.93 8.53 12.27-9.995 10.617-3.072 3.48 9.38 5.562 2.205 9.042-16.265-4.53-13.771 3.014-1.572 7.973-1.536 6.378-12.958 1.225-11.675 8.964a3.904 4.2 0 00-3.705-2.84c-7.085-2.76 4.066-3.888.578-6.338-1.807-.234-4.5 1.264-5.71-.603 9.036-6.98-11.477-.564 1.698-8.361-10.392-.584 4.952-5.173-5.205-4.609-4.916-2.314 7.663-3.656-.633-5.075-5.277-1.944 3.994-3.383.362-4.706-1.103.292-3.398-.175-2.585-1.808 5.874-1.595-.47-1.575-.632-3.889 3.759-1.167.144-1.828-.235-3.422 3.614 2.1-.38-9.392-1.934-6.631-4.862 5.833-.29 8.906-5.711 10.17 3.904 4.744-1.012 2.566-.994 5.639 3 3.441.904 3.208-2.042 3.228 3.74 6.902.361 4.783-2.946 7.136 5.422 5.211 4.301 6.144-1.988 8.07 11.53 13.32-9.543 2.955-.488 11.666a98.968 106.48 0 01-9.416 5.834c-10.627-3.54 10.609-6.359-5.241-9.586 9.036-12.62-5.007-1.187.94-12.873-2.947 0-5.025.194-2.043-3.247 0-3.053-4.862-.914-.976-5.62-4.645-1.244-3.74-.933-1.572-4.958-3.018-.117-2.91-2.937-4.157-5.212-3.163-2.139-4.464 11.356-2.169 10.034-7.012 4.608 5.025 4.22-3.867 5.736-1.374 2.198 5.873 2.411 1.03 4.59-8.314 1.419 4.283 2.76-.633 5.094-9.036-1.225 3.832 4.939-4.843 4.22-10.844-2.587 3.47-5.834-6.416-8.05 6.343-8.537-3.434-1.46.687-9.393-2.151 0-3.615.156-1.482-2.353 0-2.236-3.615-.7-.723-4.122-3.488-1.42-1.23-3.753-4.175-7.428-1.23-1.789-4.03 6.203-1.41 4.842-.307 1.206-2.892 1.614-.18 2.489-.127 1.672-4.718 1.614-.453 2.8.597 1.186-1.084 1.536-1.807 1.303-2.638.991 4.139 2.06.271 3.442-6.054 1.05 3.127 1.944-.452 3.714-6.325-.564 2.675 3.13-3.795 3.363 7.934 6.262-17.332 1.012 0-10.928-6.109-4.744-12.362.623-5.422-13.028-5.422 0-9.199.37-3.723-5.833.434-6.086-9.344-.37-1.807-10.287-12.47-2.994 1.5-1.944-10.446-18.57-5.621-4.666-2.404 7.176-5.53 8.44-3.742 1.01-.724 4.686 2.005 3.655-.74 3.014-7.229 4.006-.433 6.242-.308 4.142-11.838 4.025-1.139 6.98 1.464 2.976-2.71 3.89-4.735 3.268-6.633 2.43 10.32 5.153.669 8.614-15.164 2.586 7.825 5.017-1.139 9.275-18.597-1.05 9.488 7.35-9.524 8.42.307 4.064 8.331 1.944 8.114 6.32-3.361 5.405-17.585-1.304-5.006 8.944-1.68 4.667-14.26-2.761-11.332 4.336 5.422 1.945 8.784 5.834 2.06 8.206-13.934 2.956-9.18 24.909-18.072 35.117-9.218.117-2.82-5.736.289-9.294-4.41-3.111-8.856-.409-3.922-9.49-3.922 0-6.687.273-2.711-4.297 0-4.064-6.47-1.225-1.301-7.486-6.181-1.653-4.97-1.225-2.079-6.592-4.03-.155-3.885-3.889-5.53-6.922-4.103-3.403-1.807 5.211-4.03 6.144-2.73.74-.524 3.423 1.464 2.645-.543 2.197-5.278 2.936-.326 4.55-.217 3.014-8.62 2.936-.813 5.095 1.066 2.177-1.988 2.78-3.47 2.372-4.807 1.77 7.518 3.753.506 6.261-11.043 1.945 5.693 3.675-.831 6.767-13.519-.758 6.867 5.347-6.94 6.125 13.175 9.1 1.319 2.47-1.139 8.225 8.368 5.367-1.952 2.995-5.133 4.317 14.08-13.98-7.536-1.944 1.41-16.917-4.23 0-7.23.292-2.928-4.647.344-4.764-7.23-.292-1.41-8.05-12.65-2.684 2.223-5.095-3.614-7.584-3.506.583-5.133-14.292-7.627-4.803 1.374 3.714-6.416 7.311.307 7.292-.56 2.372-5.675 3.15-.325 4.9-.253 3.247-9.29 3.15-.904 5.483 1.157 2.334-2.132 2.995-3.614 2.567-5.205 1.945 8.096 4.025.524 6.748-11.892 2.041 6.145 3.889-.904 7.272-14.458-.817 7.428 5.833-7.464 6.611 19.14 12.25-10.844 1.109 2.44 11.959-1.807 2.683-5.711.525-8.17.875-.776 7.895 12.797 4.628-6.09 11.9-10.32.934 8.368-8.653-2.747-6.825-3.235-.447-.18-1.944-2.295-7.156-1.699-4.803-9.145-.291.56-7.564-4.229-1.653-4.934-.31-2.928-5.192-1.21-.175-4.97.856-3.162-1.789 4.102-1.73 1.265-2.819-1.555-2.741-.434-.234.524-7.409-2.891-4.317 1.698 6.845-6.019 3.539-2.278 7.136 5.187 3.209-5.042-.272 0 5.484-7.861-.37 1.808 3.558-1.355 5.833-5.223-.622.795 2.664-.29 4.395-13.301-3.248 4.067 7.136-5.584 5.93-1.807 1.945 5.241 3.112 1.247 5.056-8.53-2.605-7.464 1.945-.885 4.453-.67 3.345-9.453 1.264-5.748 5.834-5.096 1.944-14.458 2.702-7.753-5.387-7.374-1.264-5.874-17.267-10.23-.816-2.927.777-.56 3.636 1.555 2.82-.578 2.352-5.657 3.13-.343 4.86-.235 3.248-9.218 3.15-.886 5.445 1.157 2.333-2.115 2.975-3.615 2.547-5.169 1.945 8.043 4.025.525 6.709-11.802 1.944 6.108 3.889-.886 7.233-14.459-.816 7.356 5.717-7.41 6.553 18.995 12.173-10.753 1.109 2.404 11.9-1.7 2.645-5.675.506-8.115.856-4.97 3.578 10.844 5.095.813 9.003-2.837-.428-7.012 3.558-3.614 6.514-1.97 2.392-9.796 4.472-3.344.292-5.73-2.392-2.837-1.556-1.446-4.9-7.536-.098 2.783-2.742-2.078-3.656-3.615-1.575 2.837-1.225-1.446-4.18 4.012 0-.633-2.14.18-4.376-1.5-5.678-2.457 3.228-4.554 2.878-3.38 1.595 5.53 3.306-2.187 4.55 3.344 4.064-.867 2.178-.849 4.823 2.566 2.936.777 2.741-1.807 2.76 3 5.134.162 4.395-2.53 6.126 4.644 4.453 3.614 5.25-1.7 6.922 11.55 10.345-24.344 15.031-5.656-1.516-5.133-3.617-10.32-.467-4.573-11.045-4.572 0-7.771.311-3.144-5.017.361-5.133-7.916-.311-1.537-8.711-9.994-1.945.705-2.995-8.855-15.731-4.754-3.89-2.043 6.067-4.681 7.156-3.163.855-.615 3.889 1.699 3.091-.615 2.548-6.145 3.403-.362 5.29-.271 3.5-10.049 3.402-.976 5.833 1.247 2.528-2.295 3.247-4.012 2.78-5.62 2.062 8.747 4.356.578 7.292-12.85 2.198 6.633 4.259-.976 7.875-15.723-.894 8.025 6.203-8.06 7.117.253 3.442 7.048 1.634 6.868 5.367-20.025 4.764 8.35 9.373-13.067 8.517-5.422 3.889 11.711 5.833.976 9.722-35.351 31.89-72.292 78.577-122.264 77.625-462.128 73.268 547.54 34.495 544.612 43.42h0zM1290.45 213.344c3.29-2.275 8.874 4.53 1.572 3.092-1.012-.661-3.199-1.79-1.572-3.092zm-20.802-.817c2.205-1.847 1.247.35-.325.817-.778.156-.127-.37.325-.817zm-98.914 58.899a8.567 9.217 0 002.82 1.42c-1.23.272-3.56-1.537-2.82-1.42z" />
-        <path d="M798.321 469.78c9-12.25-20.477 1.421 2.425-19.942 5.337-10.78-18.402 1.249-9.506-14.014 3.262-5.537-6.733-5.072-2.669-10.56 3.803-9.016-9.732-1.396-9.07-10.045.471-7.203 9.82-3.993 10.274-11.098-2.215-3.332-15.105-1.789-6.646-10.51 1.36-1.569 1.1-4.73-.854-4.313-19.675 2.254 12.924-13.083-8.878-12.25-.122-18.178-6.977-20.31-6.977 1.054-3.75-.147-10.657 1.176-4.082 6.786 12 10.07-10.831-2.23-.157 15.337-16.884-1.102 3.872 9.923-2.895 16.22-11.198-1.765 1.744 7.472-.61 12.25-3.873-.27-16.902-5.783-11.634 5.242 2.459 2.646 9.488 4.092 7.255 10.266-24.419.857-2.947-8.06-10.849-12.642-16.517.44 5.495-11.858 3.227-18.939-16.78.662-8.145-2.23-12.732-19.918 4.936-11.662-23.966 2.278 1.203-21.143-8.948-5.61-11.18.784-6.297-14.529-20.669 1.985 9.873-13.18-10.081-12.617.261-17.37-7.36-19.33-6.977 1.078-3.75-.172-10.657 1.151-4.081 6.762 12 10.07-10.815-2.23-.157 15.361-16.884-1.127 3.89 9.8-2.896 16.194-11.18-1.764 1.744 7.497-.61 12.25-24.594-3.87-7.814 2.646-4.204 15.141-1.29 4.9-12.802-2.45-6.82 6.542 11.18 8.28-3 10.829-8.372 5.83 1.954-11.049-21.855 1.054-1.238-16.365-21.28-2.279 10.97-12.25-10.047-11.589-9.663-6.174 14.477-9.065-1.203-12.813-10.187-4.753 7.709-8.501.697-11.858-2.128.76-6.558-.392-5.006-4.508 12.053-5.071-1.256-3.479-1.186-9.628 8.582-4.214-.366-3.994-.453-8.6 7.744 3.308-3.384-34.3-6.402-7.766-1.395 5.61-9.575 2.23-2.703 9.408 4.395 9.653-4.36 3.062-5.669 7.35 8.198 14.7-2.738 5.365-1.9 14.16 6.156 9.8 1.935 7.718-3.925 8.159 4.552 9.8 3.628 20.139-6.42 14.014-2.825-4.092 7.536-9.555.699-10.878-18.803 1.053 10.849-15.63 1.552-19.6-13.57 2.818-6.663-1.274-11.04-19.306 3.784-9.016-9.751-1.42-9.07-10.045.54-6.615 8.057-4.459 9.994-9.457 1.64-6.492-16.326-1.764-6.384-12.25 4.71-8.918-11.844 1.348-6.367-8.991 8.721-4.9 2.721-7.766-3.349-7.644-.941-.661 1.1-20.727-6.192-12.054 3.053 21.805-13.36 8.722-4.848 19.918 12 10.07-10.832-2.254-.157 15.337-16.884-1.102 3.872 9.8-2.896 16.22-11.198-1.79 1.744 7.35-.61 12.25-24.611-3.872-7.814 2.645-4.204 15.165-1.43 3.822-5.32.833-7.762 1.396-3.157 4.606 14.565 15.019-1.517 13.573 13.343-13.94-17.442 1.397-.802-21.437-9.506-4.288-11.024-4.19-2.86-15.092-4.693-5.243-9.838 1.666-4.222-13.328-4.326-.319-7.483 1.225-2.913-6.052.35-6.272-7.343-.294-1.413-10.535-6.976-2.327-4.988-1.617-2.232-9.26-7.605-1.667-5.686-22.76-10.291-1.103-2.948 1.029-.558 4.9 1.57 3.748-.558 2.965-5.721 4.214-.349 6.37-.227 4.214-9.297 4.165-.89 7.154 1.169 3.063-2.128 3.92-3.715 3.357-5.232 2.45 8.093 5.267.524 8.82-12.21 3.16 6.348 4.9-.89 9.481-15.698-.44 8.337 6.91-7.465 8.624 19.186 16.44-11.128.76 2.442 15.607-1.326 4.9-11.18-2.818-8.878 4.459 4.046 2.058 7.168 5.512 1.517 8.501-7.169 2.646-3.488 4.263-5.669-1.519-11.058.613 6.402-9.236.925-11.54-9.471 2.45-3.018-4.679-5.808-8.207a1.238 1.74 0 01-.698-3.16c.418-6.787-13.43-2.083.54-11.49.96-3.847-9.627-1.054-3.75-7.18 2.774-5.267-6.976.785-3.767-5.291 5.233-2.891 1.605-4.582-1.97-4.508-.56-.392.644-12.25-3.646-7.105 1.744 13.67-8.18 4.287-2.86 11.735 6.854 5.733-6.402-1.004-.088 9.04-9.942-.66 2.267 5.856-1.744 9.556-6.593-1.054.976 4.41-.367 7.227-12.453-2.45-5.86 1.348-2.441 8.82-36.315-13.279 27.122-20.8-17.059-35.133-1.744-9.996 8.582-12.25 10.64-20.457-4.483-11.637-20.564 8.893-8.058-25.455-8.268-.613-14.303 2.45-5.564-11.564 1.744-12.863-14.495 2.033-2.704-20.09-.75-5.562-6.14 0-8.372-1.862-5.233-5.194 5.948-12.25 4.099-15.827-14.704-1.985-9.873-45.619-19.64-2.107-5.634 1.984-1.064 9.163 2.982 7.13-.104 6.418-12.785 6.345-.645 12.25 2.023 6.663-9.75 3.699-8.46 9.97.855 3.308 5.111 1.005 6.768 3.676 2.198 5.806-4.081 7.497-7.099 6.394-9.924 4.778 15.454 10.045 1.012 16.832-21.227 4.655 11.46 9.996-1.744 18.154-30.89-2.45 15.21 16.44-11.791 14.308-11.18-5.17-13.082-4.9-3.384-17.91-5.494-6.394-11.721 2.646-4.988-15.728-5.128-.368-8.861 1.445-3.489-7.154.75-7.767-8.86.637-1.744-12.422-8.442-3.16-6.297-1.127-2.651-10.951-5.128-.27-4.936-6.468-6.977-11.515-5.233-5.635-2.25 8.673-5.128 10.216-3.488 1.225-.663 5.66 1.849 4.41 0 3.97-7.901 3.92-.401 7.546 0 5.12-11.355 4.533-1.047 8.453 1.36 3.577-2.529 4.63-4.395 3.944-6.14 2.965 9.558 6.223.628 10.413-13.623 3.258 7.36 5.782-1.064 11.22-18.472-.514 9.854 8.159-8.721 10.192 18.227 15.827-1.256 4.46-1.448 13.671 15.21 11.221-10.797 1.274-6.07 10.143 13.814 8.33-5.407 11.932-4.413 1.936-11.058-.196 6.733-8.747 1.36-11.466-7.255 2.45-4.813-6.15-6.07-11.834.123-7.178-12.976-1.739.978-11.416 1.098-3.773-9.559-1.764-3.489-7.35 2.965-5.047-6.977.294-3.488-5.562 5.232-2.45 1.744-4.434-1.745-4.63-.54-.441 1.117-12.25-3.366-7.35 1.29 13.769-8.337 3.7-3.296 11.515 6.61 6.198-6.35-1.47-.454 9.016-9.89-1.372 2.058 6.002-2.058 9.407-6.54-1.518.82 4.484-.628 7.203-18.262-6.786 5.494 11.711-7.413 9.237 8.512 14.283-18.61 2.646-3.384-9.09-4.692-5.243-9.837 1.642-4.203-13.327-4.343-.319-7.5 1.225-2.93-6.052.348-6.272-7.344-.294-1.396-10.535-7.169-2.254-4.884-1.788-2.25-9.26-4.326-.221-4.169-5.464-5.948-9.8-4.413-4.754-1.9 7.35-4.343 8.648-2.947 1.053-.558 4.9 1.57 3.748-7.64 7.693 4.465 4.9-2.407 9.114-7.762 1.813 7.762 6.15-1.57 7.865.663-5.709-13.36 1.519-7.448-8.943 2.355-4.042-4.691-3.577-2.04-7.57 3.21-6.615-6.977-1.25-6.454-7.497.366-5.243 7.151-2.842 7.5-8.012-1.587-2.45-11.006-1.42-4.814-7.693 3.489-6.468-8.72.907-4.622-6.566 1.273-1.151 3.872-1.592 3.366-4.19-11.128.54-2.232-15.557-10.238-10.216-1.884 2.23 1.622 7.497-.61 9.555-2.722-.122-7.762.784-2.983 4.9 8.46 7.154-7.902-1.323-.157 11.172-12.21-.906 2.79 7.35-2.146 11.785-10.831-1.176 9.628 13.573-6.436 7.815 2.425-.93 3.053-6.296-.192-5.684-18.802 1.054 10.85-15.63 1.553-19.6-13.57 2.744-6.628-.955-11.059-19.281 4.954-11.662-23.965 2.278 1.221-21.168-8.965-5.586-11.197.809-6.296-14.504-2.6-.465-10.622 2.45-6.733-4.9 8.721-4.9 2.721-7.766-3.349-7.644-.942-.661 1.1-20.727-6.192-12.054 3.035 21.585-13.22 8.845-4.849 19.894 12 10.07-10.831-2.23-.157 15.362-16.884-1.127 3.873 9.8-2.895 16.194-11.198-1.764 1.744 7.472-.61 12.25-24.611-3.847-7.815 2.67-4.204 15.165-12.68 5.145-8.721-.882-2.738 12.52-4.535 7.742-16.536-7.742-16.413 5.904-18.332-10.265 19.325-10.51-8.425-21.045 16.727-31.605-8.93 1.078 1.605-28.396-5.128-.392-8.86 1.446-3.489-7.153.75-7.767-8.86.636-1.744-12.446-.453-3.43-3.785 0-5.233-1.152 3.384-18.179.297-2.45-4.482-21.315-5.233-5.61-2.25 8.698-5.128 10.241-3.489 1.2-.663 5.66 1.849 4.386 0 3.993-7.92 3.92-.402 7.546 0 5.17-11.354 4.532-1.046 8.452 1.448 4.9-3.488 3.773-5.512 4.214.35 7.84 11.564 5.145-2.511 12.03-4.745.195-4.797-2.45-2.687-7.84-2.755 0-4.726.244-1.9-3.97.226-4.091-4.815-.195-.925-6.884-4.553-1.69-3.489-.76-1.465-6.076-2.843-.147-2.739-3.577-3.907-6.37-1.483-3.21-5.128 10.609-1.745 8.11-.348 1.935-3.732 2.744-.209 4.165-.174 2.768-6.087 2.744-.593 4.679.768 2.009-1.395 2.45-2.424 2.205-3.489 1.617 5.232 3.43.348 5.757-7.779 1.74 4.012 3.381-.593 6.223-9.523-.71 4.867 4.9-4.883 5.635 9.227 8.06 1.064 2.058-.803 7.57 5.53 4.215-.854 3.823-3.61 3.406 9.802-9.58-17.547-5.782-9.558-17.321 17.808-15.165-19.431-8.06 2.18-28.885-5.39-7.693-13.5 4.018-5.582-17.738-5.756-.441-9.942 1.617-3.872-8.036.837-8.747-9.96.71-1.866-13.965-16.657-3.97 3.314-10.315-4.867-13.132-4.674 1.004-6.802-24.77-10.098-8.33 1.744 6.443-8.53 12.642.4 12.642 0 4.459-8.877 4.41-.435 8.477-.105 5.782-12.785 5.096-1.186 9.506 1.534 4.042-2.826 5.194-4.937 4.434-6.976 3.332 10.745 6.983.698 11.711-15.454 3.626 8.32 6.542-1.186 12.618-20.756-.588 11.076 9.163-9.907 11.417.314 5.537 8.721 2.621 8.442 8.599-3.332 7.178-18.122-1.764-5.233 12.25-1.744 6.345-14.982-3.773-11.668 6.027 5.424 2.621 9.261 7.35 1.9 11.172a9.593 13.475 0 00-5.092 2.45c-36.629-11.613 24.837-22.932-17.983-35.476-1.744-9.996 8.581-12.372 10.64-20.457-4.466-11.638-20.565 8.893-8.041-25.456-8.285-.612-14.32 2.45-5.582-11.54 1.744-12.886-14.477 2.01-2.686-20.113-.75-5.562-6.14 0-8.39-1.862-5.232-5.194 5.948-12.25 4.117-15.827-8.285-.417-7.989-10.462-11.373-18.62-8.424-9.04-3.627 14.014-8.285 16.513-5.616 1.984-1.063 9.163 3 7.13-.104 6.418-12.802 6.345-.645 12.249 2.006 6.664-9.767 3.7-8.477 9.971 7.483 2.18 10.343 11.417-.331 10.07-9.925 4.777 15.454 10.045 1.012 16.831-21.228 4.655 11.476 9.996-1.745 18.13-30.733-.318 16.675 12.618-14.25 16.464.454 7.938 12.471 3.749 12.21 12.25-5.582 10.829-25.291-3.626-7.623 17.762-29.233-6.786 11.094-14.895-15.035-22.956 16.745-31.63-8.913 1.054 1.605-28.42-5.11-.392-8.843 1.446-3.488-7.154.732-7.766-8.879.637-1.745-12.421-.47-3.455-3.802 0-5.232-1.152 3.383-18.228.366-2.45-4.483-21.315-5.233-5.634-2.25 8.673-5.11 10.217-3.489 1.2-.68 5.66 1.744 4.41 0 3.969-7.919 3.895-.384 7.546-.105 5.145-11.355 4.508-1.064 8.452 1.465 4.9-3.488 3.773-5.494 4.19-.715 7.13 12.21 5.292-2.093 11.588 6.68-7.35-5.756-7.35-1.622-13.793 5.372-10.315-21.716-2.01-4.954-16.856 2.53-.956 6.157-2.01 6.28-6.223-1.24-4.582-11.861.784-9.96-8.06 5.982-6.321 3.052-9.482-3.227-7.203-1.919-3.185-.331 295.002-.401 378.522H814.21c10.326-11.05 6.489-25.455-15.89-44.1h0zM17.123 195.382c-2.372.27-5.32-2.18-5.232-5.512 1.046 2.107 6.401 1.176 5.232 5.512zm187.451 44.957c.21 0 .384.245.576.343 4.465 5.072-5.826 4.508-.576-.367v.024zm-3.314-6.541c-.558-3.038 2.302-4.043 3.28-6.125 10.377 4.704 2.912 10.56-3.28 6.1v.025zm134.514 77.811c.209-1.788 2.825-5.439 3.645-2.989s-2.163 3.724-3.645 2.965v.024zm-.855-5.83c-6.035-1.446-.087-7.694 2.581-9.458 9.733 4.41-2.86 4.974-2.581 9.433v.024zm114.787 41.796a4.448 6.247 0 011.308 1.054c-.628 0-1.221-.05-1.308-1.078v.024zm4.133-9.8c-2.755-2.45-6.662-2.45-9.523-4.336-8.372-18.375 27.733-1.054 9.523 4.336zm226.295 90.356a1.221 1.715 0 01.715-2.083c-.61.931.14 4.533-.715 2.083zm68.896 28.224h2.616c-.261 1.543-1.744.024-2.581.073l-.035-.073h0z" />
+      <g>
+        <g stroke="null" id="svg_1" fill="url(#prefix__a)">
+          <path
+            stroke="null"
+            id="svg_2"
+            d="m1978.36511,521.32098c-14.22375,-0.37 27.49463,-355.00601 -18.05212,-315.008c-3.29895,-5.67799 24.87561,-12.97 0.85535,-11.414c-10.67542,-0.856 2.5188,-7.778 -4.32983,-9.334c-6.26904,-0.81599 4.68237,-8.42 -5.53894,-6.591c-22.00586,-2.625 18.63159,-10.84999 0.32727,-11.53099c-2.92139,0.272 -6.69812,-1.614 -5.41357,-3.636c14.1499,-8.673 -11.38062,-0.603 -4.55688,-7.778c10.22119,-3.092 3.14709,-5.037 -3.90222,-4.90001c-1.08386,-0.42799 1.28308,-13.28099 -7.20251,-7.778c3.72668,14.408 -16.01245,4.97701 -5.66431,12.755c12.94202,5.834 -12.58838,-0.48599 -0.20068,9.84c-19.66394,-0.7 4.53186,6.339 -3.34912,10.38301c-13.0675,-1.12801 1.93933,4.784 -0.72998,7.875c-36.70874,-5.93001 11.88342,12.26999 -13.92419,10.61699c-4.27979,3.48001 13.0675,5.56201 3.07178,9.04201c-22.65918,-4.53 -19.18469,3.01399 -2.19006,7.97299c-2.13989,6.37801 -18.05212,1.22501 -16.26465,8.964a5.43876,4.2 0 0 0 -5.16162,-2.84c-9.87024,-2.75999 5.66455,-3.888 0.8053,-6.338c-2.51746,-0.23401 -6.26917,1.26399 -7.95471,-0.603c12.58826,-6.98001 -15.98901,-0.56401 2.36548,-8.36101c-14.47729,-0.584 6.8988,-5.173 -7.2511,-4.60899c-6.84863,-2.31401 10.67542,-3.65601 -0.88196,-5.07501c-7.35144,-1.944 5.56421,-3.383 0.50439,-4.70599c-1.53662,0.29201 -4.73389,-0.175 -3.60132,-1.808c8.18323,-1.595 -0.65479,-1.575 -0.88037,-3.88901c5.23682,-1.16699 0.20068,-1.82799 -0.32739,-3.422c5.03479,2.10001 -0.52942,-9.392 -2.69421,-6.631c-6.77356,5.83299 -0.40405,8.90601 -7.9563,10.17c5.43884,4.744 -1.40979,2.56599 -1.38477,5.63901c4.17932,3.44099 1.25952,3.20799 -2.84473,3.228c5.21021,6.90199 0.50281,4.783 -4.10425,7.136c7.55371,5.211 5.99194,6.144 -2.76941,8.06999c16.06274,13.32001 -13.29456,2.955 -0.67993,11.666a137.87473,106.48 0 0 1 -13.11768,5.834c-14.80457,-3.53999 14.77966,-6.35899 -7.30127,-9.586c12.58826,-12.62 -6.97534,-1.187 1.30945,-12.873c-4.10535,0 -7.00024,0.194 -2.84607,-3.24699c0,-3.05301 -6.77332,-0.914 -1.35962,-5.62c-6.47107,-1.244 -5.21033,-0.93301 -2.19006,-4.95801c-4.20435,-0.117 -4.05408,-2.937 -5.79114,-5.21201c-4.40649,-2.13899 -6.21887,11.356 -3.02185,10.03401c-9.76855,4.60799 7.00049,4.22 -5.38708,5.73599c-1.91418,2.198 8.18188,2.411 1.43494,4.59c-11.58252,1.41901 5.96667,2.76001 -0.88196,5.09401c-12.58826,-1.22501 5.3385,4.939 -6.74695,4.22c-15.10693,-2.58701 4.83435,-5.834 -8.93823,-8.05c8.83655,-8.537 -4.78394,-1.46001 0.95703,-9.39301c-2.99658,0 -5.03613,0.15601 -2.06445,-2.353c0,-2.23601 -5.03613,-0.7 -1.00732,-4.12199c-4.85925,-1.42001 -1.7135,-3.75301 -5.81628,-7.42801c-1.7135,-1.789 -5.61414,6.203 -1.96423,4.84201c-0.42773,1.20599 -4.02905,1.614 -0.25085,2.489c-0.17688,1.672 -6.57275,1.614 -0.6311,2.79999c0.83179,1.186 -1.51013,1.53601 -2.51733,1.30301c-3.67493,0.991 5.76624,2.06 0.37756,3.442c-8.43396,1.05 4.35632,1.944 -0.62976,3.71399c-8.8114,-0.564 3.72668,3.13 -5.28674,3.36301c11.05298,6.26199 -24.14575,1.01199 0,-10.92799c-8.51062,-4.744 -17.22192,0.623 -7.55359,-13.028c-7.55347,0 -12.81531,0.37 -5.18665,-5.83301c0.60461,-6.086 -13.01733,-0.37 -2.51733,-10.287c-17.37219,-2.994 2.08972,-1.944 -14.55261,-18.56999c-7.83081,-4.666 -3.349,7.17599 -7.70386,8.44c-5.21326,1.00999 -1.00867,4.68599 2.79321,3.655c-1.03113,3.01399 -10.07092,4.006 -0.60339,6.24199c-0.42908,4.142 -16.4917,4.02501 -1.58667,6.98001c2.03955,2.976 -3.77551,3.89 -6.59644,3.26799c-9.2406,2.43001 14.37695,5.153 0.93201,8.614c-21.12537,2.58601 10.90112,5.01701 -1.58691,9.27501c-25.90784,-1.05 13.21802,7.34999 -13.26794,8.42c0.42761,4.064 11.60596,1.944 11.30383,6.31999c-4.6825,5.40501 -24.49829,-1.30399 -6.97412,8.944c-2.34033,4.66701 -19.86584,-2.76099 -15.78674,4.33601c7.55347,1.94499 12.23706,5.834 2.86975,8.20599c-19.41174,2.95599 -12.78882,24.909 -25.17664,35.11699c-12.84167,0.117 -3.92847,-5.73599 0.40271,-9.29399c-6.14368,-3.11099 -12.3374,-0.409 -5.46387,-9.49001c-5.46375,0 -9.3158,0.27301 -3.77686,-4.297c0,-4.064 -9.01343,-1.22499 -1.81226,-7.48599c-8.61096,-1.653 -6.92395,-1.22501 -2.89648,-6.59201c-5.61414,-0.155 -5.41211,-3.88899 -7.70386,-6.922c-5.71594,-3.403 -2.51733,5.211 -5.61426,6.144c-3.80322,0.74001 -0.7301,3.423 2.03943,2.645c-0.75635,2.19701 -7.35278,2.936 -0.45422,4.55c-0.30212,3.01399 -12.00867,2.936 -1.13257,5.095c1.48511,2.177 -2.76941,2.78 -4.83411,2.37199c-6.69678,1.77 10.47363,3.75301 0.70496,6.261c-15.38428,1.94501 7.93103,3.675 -1.15759,6.767c-18.83374,-0.758 9.56653,5.347 -9.66846,6.12502c18.35449,9.09998 1.83752,2.47 -1.58655,8.22498c11.65747,5.367 -2.71948,2.99503 -7.151,4.31702c19.61511,-13.98001 -10.49866,-1.944 1.96436,-16.91701c-5.89294,0 -10.07227,0.29201 -4.0791,-4.647c0.47925,-4.76399 -10.07239,-0.29199 -1.96436,-8.04999c-17.62305,-2.68401 3.09692,-5.095 -5.03479,-7.584c-4.88428,0.58299 -7.15088,-14.29201 -10.62524,-4.80301c1.91406,3.714 -8.93835,7.311 0.42749,7.29201c-0.78003,2.37199 -7.90588,3.14999 -0.45264,4.89999c-0.35254,3.24701 -12.94226,3.15001 -1.25952,5.483c1.61206,2.334 -2.97009,2.995 -5.03455,2.567c-7.25134,1.94499 11.27869,4.02499 0.72986,6.748c-16.56702,2.041 8.56079,3.88899 -1.25928,7.27199c-20.14185,-0.81699 10.34814,5.83301 -10.39819,6.61102c26.66418,12.25 -15.10706,1.10898 3.39905,11.95898c-2.51733,2.68301 -7.95605,0.52499 -11.38184,0.875c-1.08105,7.89499 17.82776,4.62799 -8.48413,11.89999c-14.37695,0.93402 11.65771,-8.65298 -3.82678,-6.82498c-4.50671,-0.44702 -0.25085,-1.944 -3.19727,-7.15601c-2.36694,-4.80301 -12.74023,-0.29102 0.78015,-7.564c-5.89148,-1.65302 -6.87378,-0.31 -4.07922,-5.19202c-1.68555,-0.17499 -6.92371,0.85602 -4.40503,-1.789c5.71472,-1.72998 1.76233,-2.819 -2.16614,-2.741c-0.60474,-0.23398 0.72986,-7.409 -4.02759,-4.31699c2.3656,6.845 -8.38525,3.539 -3.17358,7.13599c7.2262,3.20901 -7.02417,-0.272 0,5.48401c-10.95129,-0.37 2.51868,3.55798 -1.8877,5.83301c-7.27625,-0.62201 1.10767,2.664 -0.40405,4.39499c-18.52979,-3.24799 5.66589,7.13599 -7.77917,5.92999c-2.51733,1.94501 7.30151,3.112 1.7373,5.056c-11.8833,-2.60498 -10.39819,1.94501 -1.23291,4.453c-0.93347,3.345 -13.16919,1.26401 -8.00769,5.83401c-7.09937,1.944 -20.14185,2.702 -10.80078,-5.38702c-10.27295,-1.26398 -8.18335,-17.267 -14.25183,-0.81598c-4.07776,0.77701 -0.78003,3.63599 2.16638,2.82001c-0.8053,2.35199 -7.88086,3.12997 -0.47791,4.85999c-0.32739,3.24802 -12.8418,3.14999 -1.23425,5.44501c1.61182,2.33301 -2.94641,2.97501 -5.03613,2.547c-7.20105,1.94501 11.20483,4.02499 0.73145,6.70898c-16.44165,1.944 8.50916,3.88901 -1.23438,7.233c-20.14307,-0.81598 10.2478,5.71701 -10.323,6.55301c26.46228,12.173 -14.98035,1.10901 3.34912,11.89999c-2.36841,2.64502 -7.90613,0.50601 -11.30542,0.85602c-6.92371,3.578 15.10706,5.095 1.13281,9.00299c-3.95239,-0.42801 -9.76868,3.55801 -5.03479,6.51401c-2.74463,2.392 -13.64709,4.47198 -4.65857,0.29199c-7.98279,-2.392 -3.95239,-1.556 -2.01453,-4.89999c-10.49866,-0.09799 3.87695,-2.742 -2.8949,-3.65601c-5.03613,-1.57501 3.95227,-1.22501 -2.01453,-4.17999c5.58936,0 -0.88184,-2.14001 0.25085,-4.37601c-2.08972,-5.67801 -3.42297,3.228 -6.34424,2.87799c-4.70886,1.595 7.70386,3.306 -3.04688,4.55002c4.65869,4.064 -1.20789,2.17798 -1.18274,4.823c3.57471,2.936 1.08252,2.741 -2.51733,2.75998c4.17932,5.134 0.22571,4.39502 -3.52466,6.12601c6.46973,4.453 5.03479,5.25 -2.36829,6.922c16.09058,10.345 -33.91431,15.03101 -7.87952,-1.51599c-7.151,-3.617 -14.37708,-0.46701 -6.37073,-11.04501c-6.36938,0 -10.82593,0.311 -4.38,-5.017c0.50281,-5.133 -11.02808,-0.31097 -2.14124,-8.711c-13.92297,-1.94498 0.98206,-2.995 -12.33618,-15.73099c-6.62292,-3.89001 -2.84607,6.06699 -6.52112,7.15601c-4.40649,0.85498 -0.85681,3.88898 2.36682,3.091c-0.85669,2.54797 -8.56067,3.40298 -0.50427,5.28998c-0.37756,3.5 -13.99951,3.40201 -1.35974,5.83301c1.7373,2.52802 -3.19714,3.24701 -5.58911,2.78c-7.82935,2.06201 12.18567,4.35602 0.80518,7.29199c-17.90173,2.198 9.2406,4.259 -1.35974,7.875c-21.90405,-0.89398 11.17993,6.203 -11.22852,7.117c0.35242,3.44202 9.81873,1.634 9.56787,5.367c-27.89722,4.76401 11.63269,9.37302 -18.20386,8.517c-7.55359,3.88901 16.31494,5.83301 1.35962,9.72202c-49.24829,31.88998 -100.71167,78.577 -170.32886,77.625c-643.8017,73.26801 762.79114,34.495 758.71216,43.41998l0,0zm-179.74512,-306.97699c4.58337,-2.27501 12.36255,4.53 2.18994,3.092c-1.40979,-0.661 -4.45654,-1.78999 -2.18994,-3.092zm-28.97986,-0.817c3.07178,-1.847 1.73718,0.34999 -0.45264,0.817c-1.08398,0.15599 -0.17712,-0.37001 0.45264,-0.817zm-137.79944,58.899a11.9349,9.217 0 0 0 3.92859,1.41998c-1.7135,0.272 -4.95959,-1.53699 -3.92859,-1.41998z"
+          />
+          <path
+            stroke="null"
+            id="svg_3"
+            d="m1113.02319,470.78c12.53809,-12.25 -28.52698,1.42099 3.3783,-19.94199c7.43518,-10.78 -25.63623,1.24899 -13.24304,-14.01401c4.54431,-5.53702 -9.38,-5.07199 -3.71826,-10.56c5.29797,-9.01602 -13.55786,-1.396 -12.63562,-10.04501c0.65613,-7.203 13.68042,-3.99298 14.31287,-11.09799c-3.08569,-3.332 -21.04309,-1.789 -9.25867,-10.51001c1.89465,-1.569 1.53247,-4.72998 -1.1897,-4.31299c-27.40967,2.254 18.00476,-13.08301 -12.36816,-12.25c-0.16992,-18.17801 -9.71985,-20.31 -9.71985,1.05399c-5.22412,-0.147 -14.84644,1.17603 -5.68677,6.78601c16.71753,10.07001 -15.08887,-2.23001 -0.21863,15.33701c-23.52148,-1.10202 5.39417,9.923 -4.0332,16.22c-15.60022,-1.76501 2.42969,7.47198 -0.84973,12.25c-5.39551,-0.27002 -23.54651,-5.78302 -16.20752,5.242c3.42566,2.646 13.2179,4.09198 10.10706,10.26599c-34.01868,0.85699 -4.10559,-8.06 -15.11401,-12.642c-23.01025,0.44 7.65527,-11.858 4.49561,-18.939c-23.37665,0.66199 -11.34705,-2.23001 -17.73724,-19.918c6.8764,-11.66202 -33.38763,2.27798 1.67596,-21.14301c-12.4657,-5.61002 -15.57513,0.784 -8.77252,-14.52899c-28.79449,1.98499 13.75433,-13.18002 -14.04407,-12.617c0.36359,-17.37 -10.25348,-19.33002 -9.71991,1.078c-5.22424,-0.172 -14.8465,1.151 -5.68536,6.76199c16.71753,10.07001 -15.06659,-2.23001 -0.21869,15.36099c-23.52142,-1.12698 5.41931,9.79999 -4.03448,16.194c-15.57507,-1.76401 2.42969,7.49701 -0.84973,12.25c-34.26251,-3.87 -10.88593,2.646 -5.85675,15.14099c-1.79712,4.90002 -17.83484,-2.44998 -9.50104,6.54202c15.57507,8.28 -4.17938,10.82898 -11.66327,5.82999c2.72211,-11.04901 -30.44678,1.05402 -1.72473,-16.36499c-29.64563,-2.27902 15.28259,-12.25 -13.9967,-11.58902c-13.46173,-6.17398 20.16821,-9.065 -1.67596,-12.81299c-14.19177,-4.75299 10.73969,-8.50101 0.97107,-11.858c-2.9646,0.76001 -9.13611,-0.392 -6.974,-4.508c16.79132,-5.07101 -1.74982,-3.479 -1.65228,-9.62799c11.95587,-4.21402 -0.50983,-3.99402 -0.6311,-8.60001c10.78839,3.30798 -4.71429,-34.30002 -8.9187,-7.76599c-1.94348,5.60999 -13.33923,2.22998 -3.76563,9.40799c6.1228,9.65298 -6.07404,3.06198 -7.89764,7.35001c11.42084,14.69998 -3.81433,5.36499 -2.64697,14.16c8.57605,9.79999 2.69568,7.71799 -5.46802,8.159c6.34155,9.79999 5.05426,20.13898 -8.94379,14.01398c-3.93561,-4.09198 10.4986,-9.55499 0.97375,-10.87799c-26.19489,1.05301 15.11395,-15.63 2.16211,-19.60001c-18.90472,2.81802 -9.28235,-1.27399 -15.38007,-19.306c5.27161,-9.01599 -13.58429,-1.41998 -12.63562,-10.04498c0.75226,-6.61502 11.22437,-4.45901 13.92291,-9.457c2.28473,-6.492 -22.74414,-1.76401 -8.8938,-12.25c6.56165,-8.918 -16.50012,1.34799 -8.87,-8.991c12.14941,-4.90002 3.79071,-7.76602 -4.66559,-7.64401c-1.31091,-0.66101 1.53247,-20.72699 -8.62616,-12.05399c4.25317,21.80499 -18.61212,8.72198 -6.75391,19.918c16.71753,10.06998 -15.09027,-2.254 -0.21869,15.33698c-23.52155,-1.10199 5.39417,9.80002 -4.03448,16.22c-15.60022,-1.78998 2.42963,7.35001 -0.84979,12.25c-34.28619,-3.87198 -10.88593,2.64502 -5.85675,15.16501c-1.99213,3.82199 -7.41144,0.83301 -10.81348,1.396c-4.39807,4.60602 20.29089,15.01901 -2.11328,13.573c18.58844,-13.94 -24.29889,1.397 -1.11731,-21.43698c-13.24298,-4.28799 -15.35779,-4.19 -3.98431,-15.09201c-6.5379,-5.24301 -13.70557,1.66602 -5.88184,-13.328c-6.02661,-0.319 -10.42468,1.22501 -4.05811,-6.052c0.48755,-6.272 -10.22974,-0.29401 -1.96851,-10.535c-9.71838,-2.327 -6.94891,-1.617 -3.1095,-9.25998c-10.59467,-1.66702 -7.92126,-22.76001 -14.33661,-1.103c-4.10693,1.02899 -0.77734,4.89999 2.18719,3.74799c-0.77734,2.965 -7.97009,4.21399 -0.48621,6.37c-0.31622,4.21402 -12.95184,4.16501 -1.23981,7.15399c1.62848,3.06302 -2.96466,3.92001 -5.17548,3.35703c-7.28882,2.44998 11.27454,5.267 0.72998,8.81998c-17.01001,3.16 8.84357,4.90002 -1.23993,9.48102c-21.86926,-0.44 11.6145,6.91 -10.3996,8.62399c26.72839,16.44 -15.50269,0.76001 3.40198,15.60699c-1.84735,4.89999 -15.57507,-2.81799 -12.36823,4.45901c5.63666,2.05798 9.98596,5.51199 2.1134,8.50098c-9.9873,2.64603 -4.85913,4.263 -7.89758,-1.51898c-15.40515,0.61301 8.91876,-9.23599 1.28864,-11.54001c-13.19434,2.45001 -4.20447,-4.67899 -8.09125,-8.207a1.72469,1.74 0 0 1 -0.97241,-3.16c0.58234,-6.78699 -18.70966,-2.08298 0.75226,-11.48999c1.3374,-3.84702 -13.41156,-1.05399 -5.22424,-7.17999c3.86462,-5.267 -9.71832,0.785 -5.24786,-5.29102c7.29022,-2.89099 2.23596,-4.582 -2.74445,-4.508c-0.78015,-0.392 0.89722,-12.25 -5.07935,-7.10501c2.42963,13.67001 -11.39575,4.28702 -3.98431,11.73502c9.54846,5.733 -8.91876,-1.004 -0.12262,9.04001c-13.8504,-0.66 3.15826,5.85599 -2.42957,9.556c-9.18488,-1.05402 1.35962,4.41 -0.51129,7.22699c-17.34857,-2.45001 -8.1637,1.34799 -3.40063,8.82001c-50.59131,-13.27899 37.7843,-20.80002 -23.76532,-35.133c-2.42957,-9.996 11.95581,-12.25 14.82288,-20.457c-6.24536,-11.63702 -28.64819,8.89297 -11.22583,-25.45502c-11.51837,-0.61301 -19.92584,2.45001 -7.75134,-11.564c2.42963,-12.86301 -20.1933,2.03299 -3.76703,-20.09c-1.0448,-5.562 -8.55371,0 -11.66321,-1.862c-7.29022,-5.194 8.28632,-12.25 5.71045,-15.827c-20.4845,-1.985 -13.75433,-45.619 -27.36096,-2.10701c-7.84888,1.98401 -1.4823,9.16301 4.15424,7.13c-0.14484,6.418 -17.81104,6.345 -0.8985,12.25c2.8183,6.66301 -13.58301,3.69901 -11.78583,9.96999c1.1911,3.30801 7.12024,1.005 9.42865,3.67603c3.06207,5.806 -5.68536,7.49698 -9.88977,6.39398c-13.82538,4.77802 21.52936,10.04501 1.40985,16.832c-29.57184,4.655 15.96515,9.996 -2.42969,18.15399c-43.03357,-2.44998 21.18945,16.44 -16.42627,14.30801c-15.57513,-5.17001 -18.22485,-4.89999 -4.71436,-17.91c-7.65381,-6.39398 -16.3288,2.646 -6.94891,-15.728c-7.14392,-0.36801 -12.34448,1.44501 -4.8606,-7.15399c1.04486,-7.767 -12.34308,0.63699 -2.42963,-12.422c-11.76074,-3.16 -8.77246,-1.12701 -3.69318,-10.951c-7.14392,-0.27 -6.87646,-6.468 -9.71979,-11.51501c-7.29022,-5.63499 -3.13452,8.673 -7.14392,10.216c-4.85925,1.22501 -0.92365,5.66002 2.57587,4.41002c0,3.96997 -11.00708,3.91998 -0.55865,7.54599c0,5.12 -15.81891,4.53299 -1.45862,8.453c1.89465,3.577 -3.52319,4.63 -6.1228,3.944c-8.55371,2.965 13.31555,6.22299 0.87494,10.41299c-18.97858,3.258 10.25336,5.78198 -1.4823,11.22c-25.73376,-0.51401 13.72784,8.159 -12.14941,10.19199c25.39246,15.827 -1.74982,4.46002 -2.01727,13.67102c21.18945,11.22098 -15.04156,1.27399 -8.45624,10.14297c19.24457,8.33002 -7.53265,11.93201 -6.14789,1.936c-15.40515,-0.19598 9.37994,-8.74698 1.89465,-11.466c-10.10712,2.45001 -6.70508,-6.14999 -8.4563,-11.83398c0.17139,-7.17801 -18.07715,-1.73901 1.36255,-11.41602c1.5296,-3.77298 -13.31689,-1.76398 -4.86066,-7.34998c4.13062,-5.047 -9.71979,0.29398 -4.85919,-5.56201c7.28882,-2.45001 2.42963,-4.43399 -2.43103,-4.63c-0.75226,-0.44101 1.55615,-12.25 -4.68927,-7.35001c1.79718,13.76901 -11.61444,3.70001 -4.59167,11.51501c9.2085,6.198 -8.84637,-1.47 -0.63251,9.01599c-13.77802,-1.37201 2.86707,6.00201 -2.86707,9.40701c-9.11102,-1.51801 1.14233,4.48398 -0.87488,7.203c-25.44122,-6.78601 7.65387,11.711 -10.32721,9.237c11.85828,14.28299 -25.92609,2.646 -4.71436,-9.09c-6.5365,-5.24301 -13.70416,1.642 -5.85529,-13.32703c-6.05032,-0.319 -10.44846,1.22501 -4.08185,-6.052c0.4848,-6.27197 -10.23111,-0.29398 -1.94479,-10.53497c-9.98734,-2.254 -6.80405,-1.78802 -3.13455,-9.26001c-6.02667,-0.22101 -5.80792,-5.46399 -8.28629,-9.79999c-6.14789,-4.75403 -2.64697,7.34998 -6.05035,8.64798c-4.10553,1.05301 -0.77734,4.89999 2.18719,3.74802c-10.64349,7.69299 6.22031,4.89999 -3.35327,9.11398c-10.81342,1.81302 10.81348,6.14999 -2.18716,7.86502c0.92365,-5.70901 -18.61215,1.51898 -10.37598,-8.94302c3.28076,-4.04199 -6.53516,-3.577 -2.84198,-7.56998c4.47189,-6.61502 -9.71982,-1.25 -8.99124,-7.49701c0.50989,-5.24301 9.96222,-2.84201 10.44843,-8.01199c-2.21091,-2.45001 -15.3327,-1.42001 -6.70648,-7.69301c4.86063,-6.468 -12.14807,0.907 -6.43906,-6.56601c1.77347,-1.15099 5.3942,-1.592 4.68927,-4.18999c-15.50269,0.53999 -3.10947,-15.55701 -14.26279,-10.216c-2.62463,2.23 2.25961,7.49699 -0.84982,9.55499c-3.79205,-0.12199 -10.81342,0.78401 -4.1557,4.90001c11.78589,7.15399 -11.00842,-1.323 -0.21869,11.17198c-17.01007,-0.90598 3.88678,7.35001 -2.98962,11.785c-15.08896,-1.17599 13.41296,13.573 -8.96616,7.815c3.3783,-0.92999 4.25317,-6.29599 -0.26752,-5.68399c-26.19351,1.05399 15.11542,-15.63 2.16354,-19.60001c-18.90472,2.74399 -9.23361,-0.955 -15.40656,-19.28099c6.90152,-11.662 -33.3862,2.278 1.70099,-21.168c-12.48935,-5.586 -15.59882,0.80899 -8.77112,-14.50401c-3.62213,-0.465 -14.79776,2.45001 -9.37988,-4.89999c12.14944,-4.89999 3.79068,-7.76599 -4.66559,-7.644c-1.31229,-0.661 1.53244,-20.72701 -8.62619,-12.054c4.22809,21.58501 -18.41711,8.845 -6.75528,19.894c16.7175,10.07001 -15.08893,-2.23 -0.21872,15.362c-23.52151,-1.127 5.39554,9.8 -4.03308,16.194c-15.60022,-1.76399 2.4296,7.472 -0.84982,12.25c-34.28619,-3.847 -10.88727,2.67 -5.85672,15.16501c-17.66479,5.145 -12.14941,-0.882 -3.81436,12.52c-6.31781,7.742 -23.03668,-7.742 -22.86536,5.90399c-25.53873,-10.26501 26.92215,-10.51001 -11.73706,-21.045c23.3028,-31.60501 -12.44058,1.078 2.23599,-28.39601c-7.14395,-0.392 -12.34311,1.44601 -4.86066,-7.153c1.04486,-7.767 -12.34308,0.636 -2.42957,-12.446c-0.6311,-3.42999 -5.27298,0 -7.29022,-1.15199c4.71432,-18.179 0.41373,-2.45 -6.24402,-21.315c-7.29019,-5.61 -3.13452,8.698 -7.14392,10.241c-4.8606,1.2 -0.92365,5.66 2.5759,4.386c0,3.993 -11.03354,3.92 -0.56006,7.54601c0,5.17 -15.8175,4.532 -1.45718,8.452c2.01724,4.90001 -4.85922,3.77299 -7.67892,4.214c0.48761,7.84 16.11011,5.145 -3.49814,12.03c-6.61035,0.19499 -6.6828,-2.45 -3.74332,-7.84c-3.83807,0 -6.58389,0.24399 -2.64691,-3.97c0.31482,-4.091 -6.70792,-0.19501 -1.28867,-6.884c-6.34286,-1.69 -4.8606,-0.75999 -2.04092,-6.076c-3.96063,-0.147 -3.81577,-3.577 -5.44293,-6.37c-2.06601,-3.21001 -7.14395,10.60899 -2.431,8.11c-0.4848,1.935 -5.19913,2.744 -0.29117,4.16499c-0.2424,2.76801 -8.47995,2.744 -0.82614,4.679c1.06995,2.009 -1.94339,2.45 -3.37692,2.205c-4.8606,1.617 7.28882,3.43001 0.48483,5.757c-10.83713,1.73999 5.5892,3.381 -0.82614,6.22299c-13.26672,-0.70999 6.78036,4.90001 -6.80261,5.63501c12.85434,8.06 1.48227,2.058 -1.11868,7.56999c7.70398,4.215 -1.18974,3.823 -5.02919,3.40601c13.65541,-9.58 -24.44514,-5.782 -13.31548,-17.321c24.80878,-15.16501 -27.06979,-8.06 3.03702,-28.88501c-7.50896,-7.69299 -18.80717,4.01801 -7.77641,-17.73799c-8.01884,-0.44101 -13.85045,1.617 -5.39417,-8.036c1.16603,-8.74701 -13.87553,0.70999 -2.59958,-13.96501c-23.20526,-3.97 4.61681,-10.315 -6.78033,-13.13199c-6.51146,1.004 -9.47606,-24.77 -14.06778,-8.33c2.42961,6.44299 -11.88335,12.642 0.55727,12.642c0,4.459 -12.36676,4.41 -0.606,8.47699c-0.1463,5.78201 -17.81111,5.09601 -1.65225,9.50601c2.13704,4.04199 -3.93698,5.194 -6.87785,4.43399c-9.71843,3.332 14.96912,6.983 0.9724,11.711c-21.52936,3.62601 11.59077,6.54201 -1.65225,12.61801c-28.9157,-0.58801 15.43025,9.16299 -13.80168,11.41699c0.43744,5.537 12.14943,2.621 11.76076,8.599c-4.64189,7.17801 -25.2462,-1.76399 -7.29022,12.25c-2.42961,6.345 -20.8718,-3.77299 -16.25497,6.02701c7.55632,2.621 12.90173,7.34999 2.64694,11.172a13.36424,13.475 0 0 0 -7.0938,2.45c-51.02876,-11.61299 34.60104,-22.93199 -25.05255,-35.476c-2.42961,-9.996 11.95441,-12.37199 14.82285,-20.457c-6.22169,-11.638 -28.6496,8.89301 -11.2021,-25.45599c-11.54204,-0.612 -19.94955,2.45 -7.77643,-11.54001c2.42961,-12.886 -20.16827,2.01001 -3.74193,-20.11299c-1.04485,-5.562 -8.55379,0 -11.68832,-1.862c-7.28883,-5.19401 8.28629,-12.25001 5.7355,-15.827c-11.54204,-0.417 -11.12968,-10.46201 -15.844,-18.62c-11.73569,-9.04 -5.05286,14.014 -11.54204,16.513c-7.82378,1.984 -1.48089,9.163 4.17937,7.13c-0.14487,6.41801 -17.83477,6.34501 -0.89856,12.249c2.7946,6.664 -13.60664,3.7 -11.80952,9.97101c10.42475,2.17999 14.40909,11.41699 -0.46112,10.06999c-13.82676,4.77701 21.52934,10.045 1.40984,16.83099c-29.57324,4.65501 15.9875,9.996 -2.43099,18.12999c-42.81489,-0.31799 23.23034,12.61801 -19.85202,16.464c0.63248,7.938 17.37364,3.74901 17.01004,12.25c-7.77642,10.82901 -35.2335,-3.62599 -10.61978,17.76201c-40.7252,-6.78601 15.45532,-14.895 -20.94563,-22.95601c23.32787,-31.62999 -12.41692,1.054 2.23596,-28.42c-7.11887,-0.392 -12.31939,1.446 -4.85921,-7.15399c1.01976,-7.76601 -12.36955,0.63699 -2.431,-12.42099c-0.65477,-3.455 -5.29666,0 -7.28883,-1.15201c4.71294,-18.22801 0.50988,-2.45 -6.24538,-21.315c-7.29022,-5.634 -3.13453,8.67299 -7.11886,10.217c-4.86061,1.2 -0.94732,5.66 2.42961,4.41c0,3.96899 -11.03215,3.89499 -0.53496,7.54599c-0.14628,5.145 -15.81892,4.50801 -1.48228,8.45201c2.04093,4.89999 -4.85922,3.77299 -7.65383,4.18999c-0.99608,7.13 17.01005,5.29201 -2.91581,11.58801c9.30607,-7.35001 -8.01882,-7.35001 -2.25965,-13.793c7.48386,-10.315 -30.25309,-2.01001 -6.90154,-16.856c3.5246,-0.95601 8.57746,-2.01001 8.74882,-6.22301c-1.72747,-4.582 -16.52385,0.784 -13.87552,-8.06c8.33367,-6.321 4.25181,-9.48199 -4.49561,-7.203c-2.67341,-3.185 -0.46112,295.00203 -0.55864,378.52202l1136.58603,0c14.38538,-11.05002 9.04004,-25.45502 -22.13684,-44.10001l0,0l0.00146,0zm-1088.30588,-274.39799c-3.30449,0.26999 -7.41142,-2.18001 -7.28882,-5.51201c1.45721,2.10701 8.91739,1.17601 7.28882,5.51201zm261.14255,44.957c0.29254,0 0.53494,0.245 0.8024,0.343c6.22034,5.07199 -8.1163,4.508 -0.8024,-0.367l0,0.024zm-4.61685,-6.541c-0.77734,-3.03801 3.207,-4.043 4.56946,-6.125c14.45648,4.70399 4.05679,10.56001 -4.56946,6.09999l0,0.02501zm187.39474,77.811c0.29114,-1.78799 3.93555,-5.439 5.07794,-2.98901s-3.01334,3.724 -5.07794,2.965l0,0.02402zm-1.19113,-5.83002c-8.4075,-1.44598 -0.12122,-7.694 3.59564,-9.45798c13.5593,4.40997 -3.98431,4.974 -3.59564,9.43298l0,0.02402l0,0.00098zm159.91257,41.79602a6.19662,6.247 0 0 1 1.8222,1.05399c-0.87488,0 -1.70105,-0.04999 -1.8222,-1.078l0,0.02402zm5.75781,-9.80002c-3.83807,-2.44998 -9.28107,-2.44998 -13.26678,-4.336c-11.66321,-18.375 38.63556,-1.05399 13.26678,4.336zm315.25702,90.35602a1.701,1.715 0 0 1 0.99603,-2.08301c-0.84973,0.931 0.19507,4.53302 -0.99603,2.08301zm95.98071,28.224l3.64441,0c-0.36365,1.543 -2.42957,0.02399 -3.5957,0.073l-0.04871,-0.073l0,0z"
+          />
+        </g>
       </g>
     </svg>
   );
