+++ packages/demo/src/pages/home/index.scss
@@ -1,6 +1,45 @@
 .pivot-design-home {
   // background-color: var(--master-bg-color);
+  position: relative;
   min-height: 600%;
+  .pivot-design-home-title {
+    position: fixed;
+    left: 0;
+    right: 0;
+    top: 0;
+    display: flex;
+    justify-content: space-between;
+    padding: 30px 20px;
+    z-index: 100;
+    .title {
+      transform: translate3d(0, -18px, 0);
+      width: 200px;
+      height: 36px;
+      font-size: 30px;
+      color: var(--master-text-color);
+    }
+    .navigation-wrapper {
+      display: flex;
+      .navigator {
+        margin: 0 20px;
+        padding: 8px 30px;
+        border: 1px solid transparent;
+        box-sizing: border-box;
+        border-radius: 20px;
+        text-decoration: none;
+        font-size: 18px;
+        font-weight: 600;
+        color: var(--link-color);
+        border: 1px solid;
+        transition: 0.5s;
+        &:hover {
+          border-color: adjust-hue(#ffa260, 45deg);
+          box-shadow: inset 0 0 0 2em adjust-hue(#ffa260, 45deg);
+          color: #000000;
+        }
+      }
+    }
+  }
   .background {
     position: fixed;
     top: 0;
@@ -13,43 +52,7 @@
     align-items: center;
     background: linear-gradient(60deg, #332051 0%, #000000 100%);
     background-attachment: fixed;
-    .pivot-design-home-title {
-      position: absolute;
-      left: 0;
-      right: 0;
-      top: 0;
-      display: flex;
-      justify-content: space-between;
-      padding: 30px 20px;
-      .title {
-        transform: translate3d(0, -18px, 0);
-        width: 200px;
-        height: 36px;
-        font-size: 30px;
-        color: var(--master-text-color);
-      }
-      .navigation-wrapper {
-        display: flex;
-        .navigator {
-          margin: 0 20px;
-          padding: 8px 30px;
-          border: 1px solid transparent;
-          box-sizing: border-box;
-          border-radius: 20px;
-          text-decoration: none;
-          font-size: 18px;
-          font-weight: 600;
-          color: var(--link-color);
-          border: 1px solid;
-          transition: 0.5s;
-          &:hover {
-            border-color: adjust-hue(#ffa260, 45deg);
-            box-shadow: inset 0 0 0 2em adjust-hue(#ffa260, 45deg);
-            color: #000000;
-          }
-        }
-      }
-    }
+
     .big-title {
       position: absolute;
       top: 50%;
@@ -84,6 +87,7 @@
       top: 0;
       bottom: 0;
       z-index: 1;
+      pointer-events: none;
       &.left_birds,
       &.right_birds {
         opacity: 0;
@@ -115,8 +119,15 @@
         }
       }
     }
-    .home-description {
-      color: var(--master-text-color);
+    .pivot-design-description {
+      color: var(--master-bg-color);
+      transform: scale(0.4);
+      opacity: 0;
+      &-header {
+        width: 100%;
+        text-align: center;
+        font-size: 54px;
+      }
     }
   }
   .home-item {
