+++ packages/demo/webpack.config.js
@@ -17,6 +17,7 @@ module.exports = {
   },
   devServer: {
     port: 8080,
+    historyApiFallback: true,
   },
   module: {
     rules: [
